fix(qov-860): Force uint32 as u64

+++ src/apis/database_instances_api.rs
@@ -774,7 +774,7 @@ pub async fn list_instances(
     organization_id: Option<&str>,
     project_id: Option<&str>,
     page: Option<i32>,
-    page_size: Option<i32>,
+    page_size: Option<u64>,
 ) -> Result<models::ScalewayPeriodRdbPeriodV1PeriodListInstancesResponse, Error<ListInstancesError>>
 {
     // add a prefix to parameters to efficiently prevent name collisions
